Remove double precision requirement.

+++ src/torchfem/io.py
@@ -38,12 +38,12 @@ def import_mesh(filename, C):
             elements += cell_block.data.tolist()
 
     if mesh.points[:, 2].any():
-        nodes = torch.from_numpy(mesh.points.astype(np.float64))
+        nodes = torch.from_numpy(mesh.points.astype(np.float32))
         forces = torch.zeros_like(nodes)
         constraints = torch.zeros_like(nodes, dtype=bool)
         return Solid(nodes, elements, forces, constraints, C)
     else:
-        nodes = torch.from_numpy(mesh.points[:, 0:2].astype(np.float64))
+        nodes = torch.from_numpy(mesh.points[:, 0:2].astype(np.float32))
         thickness = torch.ones((len(elements)))
         forces = torch.zeros_like(nodes)
         constraints = torch.zeros_like(nodes, dtype=bool)
